pyproject.toml
[project]
name = "blog-publisher"
version = "0.1.0"
description = "Web interface for publishing blog posts to GitLab"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "gitpython>=3.1.0",
    "python-gitlab>=4.0.0",
    "pydantic>=2.0.0",
    "filelock>=3.13.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "services"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
